+++ tests/conftest.py
@@ -13,10 +13,15 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-import pytest
+import os
+import threading
+from typing import Any, Generator
+
+import asyncio
+import pytest  # noqa F401 Needed to run the tests
 
 
-def pytest_addoption(parser):
+def pytest_addoption(parser: Any) -> None:
     parser.addoption(
         "--run_private_ip",
         action="store_true",
@@ -25,16 +30,46 @@ def pytest_addoption(parser):
     )
 
 
-def pytest_configure(config):
+def pytest_configure(config: Any) -> None:
     config.addinivalue_line(
         "markers", "private_ip: mark test as requiring private IP access"
     )
 
 
-def pytest_collection_modifyitems(config, items):
+def pytest_collection_modifyitems(config: Any, items: Any) -> None:
     if config.getoption("--run_private_ip"):
         return
     skip_private_ip = pytest.mark.skip(reason="need --run_private_ip option to run")
     for item in items:
         if "private_ip" in item.keywords:
             item.add_marker(skip_private_ip)
+
+
+@pytest.fixture
+def async_loop() -> Generator:
+    """
+    Creates a loop in a background thread and returns it to use for testing.
+    """
+    loop = asyncio.new_event_loop()
+    thr = threading.Thread(target=loop.run_forever)
+    thr.start()
+    yield loop
+    loop.stop()
+    thr.join()
+
+
+@pytest.fixture
+def connect_string() -> str:
+    """
+    Retrieves a valid connection string from the environment and
+    returns it.
+    """
+    try:
+        connect_string = os.environ["INSTANCE_CONNECTION_NAME"]
+    except KeyError:
+        raise KeyError(
+            "Please set environment variable 'INSTANCE_CONNECTION"
+            + "_NAME' to a valid Cloud SQL connection string."
+        )
+
+    return connect_string
